refactor: :recycle: reference pusher instance in server
